# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TrainCoach API",
    "version": "v1"
  },
  "paths": {
...
    "/api/athletes/{athleteUserId}/food": {
      "get": {
        "tags": [
          "FoodEntries"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FoodEntryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FoodEntryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FoodEntryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FoodEntries"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodEntryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodEntryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodEntryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEntryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEntryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEntryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/athletes/{athleteUserId}/food/{entryId}": {
      "delete": {
        "tags": [
          "FoodEntries"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/athletes/{athleteUserId}/heart-rate-zones": {
      "get": {
        "tags": [
          "HeartRateZones"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartRateZoneDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartRateZoneDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartRateZoneDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "HeartRateZones"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHeartRateZonesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHeartRateZonesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetHeartRateZonesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartRateZoneDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartRateZoneDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeartRateZoneDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/athletes/{athleteUserId}/hrv": {
      "get": {
        "tags": [
          "HrvMeasurements"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HrvMeasurementDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HrvMeasurementDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HrvMeasurementDto"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "HrvMeasurements"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertHrvMeasurementRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertHrvMeasurementRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertHrvMeasurementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HrvMeasurementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrvMeasurementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HrvMeasurementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/athletes/{athleteUserId}/hydration": {
      "get": {
        "tags": [
          "HydrationEntries"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HydrationEntryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HydrationEntryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HydrationEntryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "HydrationEntries"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHydrationEntryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHydrationEntryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHydrationEntryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HydrationEntryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HydrationEntryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HydrationEntryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/athletes/{athleteUserId}/hydration/{entryId}": {
      "delete": {
        "tags": [
          "HydrationEntries"
        ],
        "parameters": [
          {
            "name": "athleteUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/import/preview": {
      "post": {
        "tags": [
          "Import"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "fileType": {
                    "$ref": "#/components/schemas/ImportFileType"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                },
                "fileType": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/{importedFileId}/confirm": {
      "post": {
        "tags": [
          "Import"
        ],
        "parameters": [
          {
            "name": "importedFileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportConfirmResultDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportConfirmResultDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportConfirmResultDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateFoodEntryRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "consumedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "mealType": {
            "$ref": "#/components/schemas/MealType"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "estimatedCarbsGrams": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "estimatedProteinGrams": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "relativeToWorkoutNote": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateHydrationEntryRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "consumedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "drinkType": {
            "$ref": "#/components/schemas/HydrationDrinkType"
          },
          "volumeMilliliters": {
            "type": "number",
            "format": "double"
          },
          "containsElectrolytes": {
            "type": "boolean"
          },
          "caffeineMilligrams": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DataSource": {
        "enum": [
          "Manual",
          "FileImport",
          "Strava",
          "GarminDemoProvider",
          "MySasyDemoProvider"
        ],
        "type": "string"
      },
      "FoodEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "consumedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "mealType": {
            "$ref": "#/components/schemas/MealType"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "estimatedCarbsGrams": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "estimatedProteinGrams": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "relativeToWorkoutNote": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HeartRateZoneDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "zoneNumber": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "minBpm": {
            "type": "integer",
            "format": "int32"
          },
          "maxBpm": {
            "type": "integer",
            "format": "int32"
          },
          "minPaceSecondsPerKm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxPaceSecondsPerKm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "effectiveFromDate": {
            "type": "string",
            "format": "date"
          }
        },
        "additionalProperties": false
      },
      "HeartRateZoneInput": {
        "type": "object",
        "properties": {
          "zoneNumber": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "minBpm": {
            "type": "integer",
            "format": "int32"
          },
          "maxBpm": {
            "type": "integer",
            "format": "int32"
          },
          "minPaceSecondsPerKm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxPaceSecondsPerKm": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HrvMeasurementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "rmssdMs": {
            "type": "number",
            "format": "double"
          },
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "measuredAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HydrationDrinkType": {
        "enum": [
          "Water",
          "ElectrolyteDrink",
          "Coffee",
          "Tea",
          "Alcohol",
          "Other"
        ],
        "type": "string"
      },
      "HydrationEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "consumedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "drinkType": {
            "$ref": "#/components/schemas/HydrationDrinkType"
          },
          "volumeMilliliters": {
            "type": "number",
            "format": "double"
          },
          "containsElectrolytes": {
            "type": "boolean"
          },
          "caffeineMilligrams": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImportConfirmResultDto": {
        "type": "object",
        "properties": {
          "importedFileId": {
            "type": "string",
            "format": "uuid"
          },
          "rowsImported": {
            "type": "integer",
            "format": "int32"
          },
          "rowsSkippedDuplicate": {
            "type": "integer",
            "format": "int32"
          },
          "rowsSkippedInvalid": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ImportFileType": {
        "enum": [
          "GoogleSheetsExportCsv",
          "StandardCsvTemplate",
          "GarminActivityExport",
          "MySasyExport"
        ],
        "type": "string"
      },
      "ImportPreviewDto": {
        "type": "object",
        "properties": {
          "importedFileId": {
            "type": "string",
            "format": "uuid"
          },
          "originalFileName": {
            "type": "string",
            "nullable": true
          },
          "rowsTotal": {
            "type": "integer",
            "format": "int32"
          },
          "rowsValid": {
            "type": "integer",
            "format": "int32"
          },
          "rowsWithWarnings": {
            "type": "integer",
            "format": "int32"
          },
          "rowsWithErrors": {
            "type": "integer",
            "format": "int32"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportPreviewRowDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImportPreviewRowDto": {
        "type": "object",
        "properties": {
          "rowNumber": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "sport": {
            "$ref": "#/components/schemas/SportType"
          },
          "distanceKm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "elevationM": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "averageHeartRate": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ImportRowStatus"
          },
          "messages": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImportRowStatus": {
        "enum": [
          "Valid",
          "Warning",
          "Error",
          "DuplicateSkipped"
        ],
        "type": "string"
      },
...
      "MealType": {
        "enum": [
          "Breakfast",
          "Lunch",
          "Dinner",
          "Snack",
          "PreWorkout",
          "DuringWorkout",
          "PostWorkout"
        ],
        "type": "string"
      },
...
      "SetHeartRateZonesRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "effectiveFromDate": {
            "type": "string",
            "format": "date"
          },
          "zones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeartRateZoneInput"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SportType": {
        "enum": [
          "Running",
          "Cycling",
          "Swimming",
          "Strength",
          "CrossTraining",
          "Rest",
          "Other"
        ],
        "type": "string"
      },
...
      "UpsertHrvMeasurementRequest": {
        "type": "object",
        "properties": {
          "athleteUserId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "rmssdMs": {
            "type": "number",
            "format": "double"
          },
          "source": {
            "$ref": "#/components/schemas/DataSource"
          },
          "measuredAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Zadejte JWT access token (bez prefixu \"Bearer \").",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}